% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_self.R
\name{add_self}
\alias{add_self}
\title{Add a number to its self}
\usage{
add_self(x)
}
\arguments{
\item{x}{a numeric value to be added to its self}
}
\value{
a number added to its self
}
\description{
Takes in any numeric value and adds it to its self
}
