% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply5.R
\name{multiply5}
\alias{multiply5}
\title{Multiply by 5}
\usage{
multiply5(x)
}
\arguments{
\item{x}{a numeric value to be multiplied by 5}
}
\value{
a number 5 times larger than the number submitted
}
\description{
Takes in any numeric value and multiples it by 5
}
