% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{a numeric value to be raised to the power of 2}
}
\value{
a number submitted squared
}
\description{
Takes in any numeric value multiples it by its self
}
