% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract1.R
\name{subtract1}
\alias{subtract1}
\title{Subtract 1 from a number}
\usage{
subtract1(x)
}
\arguments{
\item{x}{a numeric value to have 1 subtracted from it}
}
\value{
a number submitted t hat has had 1 subtracted from it
}
\description{
Takes in any numeric value and subtracts 1 from it
}
